🔧 REFACTOR: Add attachments and subtasks blueprints, migrate 15 more routes
- Added attachments blueprint for file upload/download functionality
- Added subtasks blueprint for subtask management routes
- Moved contact association routes to contacts and clients blueprints
- Added admin recurring task processing route to admin blueprint
- Created proper imports and utility function organization
- Progress: 67 routes migrated total (100 → 33 remaining)
