+++ app.py
@@ -337,142 +337,7 @@ def add_security_headers(response):
 
 # Calendar route moved to views blueprint
 
-@app.route('/kanban')
-def kanban_view():
-    firm_id = session['firm_id']
-    
-    # Get filter parameters
-    work_type_filter = request.args.get('work_type')
-    priority_filter = request.args.get('priority')
-    due_filter = request.args.get('due_filter')
-    
-    # Get work types for filtering
-    work_types = WorkType.query.filter_by(firm_id=firm_id, is_active=True).all()
-    
-    # If no work type is selected and work types exist, redirect to the first one
-    if not work_type_filter and work_types:
-        return redirect(url_for('kanban_view', work_type=work_types[0].id))
-    
-    current_work_type = None
-    kanban_columns = []
-    
-    if work_type_filter:
-        current_work_type = WorkType.query.filter_by(id=work_type_filter, firm_id=firm_id).first()
-        if current_work_type:
-            # Get the template associated with this work type
-            template = Template.query.filter_by(
-                work_type_id=current_work_type.id,
-                firm_id=firm_id
-            ).first()
-            
-            if template:
-                # Use template tasks as Kanban columns
-                kanban_columns = TemplateTask.query.filter_by(
-                    template_id=template.id
-                ).order_by(TemplateTask.order.asc()).all()
-    
-    # Base query for projects (not individual tasks)
-    query = Project.query.filter_by(firm_id=firm_id, status='Active')
-    
-    # Apply work type filter
-    if work_type_filter and current_work_type:
-        query = query.filter(Project.work_type_id == current_work_type.id)
-    
-    # Apply priority filter
-    if priority_filter:
-        query = query.filter(Project.priority == priority_filter)
-    
-    # Apply due date filters
-    today = date.today()
-    if due_filter == 'overdue':
-        query = query.filter(Project.due_date < today)
-    elif due_filter == 'today':
-        query = query.filter(Project.due_date == today)
-    elif due_filter == 'this_week':
-        week_end = today + timedelta(days=7)
-        query = query.filter(Project.due_date.between(today, week_end))
-    
-    # Get all projects
-    projects = query.order_by(Project.created_at.desc()).all()
-    
-    # Organize projects by current task progress (which template task they're currently on)
-    projects_by_column = {}
-    project_counts = {}
-    
-    if kanban_columns:
-        # Initialize columns
-        for column in kanban_columns:
-            projects_by_column[column.id] = []
-            project_counts[column.id] = 0
-        
-        # Add a "Completed" column for finished projects
-        projects_by_column['completed'] = []
-        project_counts['completed'] = 0
-        
-        for project in projects:
-            # Determine which column this project belongs in based on task progress
-            project_tasks = Task.query.filter_by(project_id=project.id).all()
-            
-            if not project_tasks:
-                # No tasks yet - put in first column
-                if kanban_columns:
-                    first_column = kanban_columns[0]
-                    projects_by_column[first_column.id].append(project)
-                    project_counts[first_column.id] += 1
-            else:
-                # Find current task stage based on completion
-                all_completed = True
-                current_column_id = None
-                
-                # Go through template tasks in order to find current stage
-                for template_task in kanban_columns:
-                    # Find corresponding project task
-                    project_task = next(
-                        (t for t in project_tasks if t.template_task_origin_id == template_task.id),
-                        None
-                    )
-                    
-                    if project_task:
-                        if project_task.status != 'Completed':
-                            # This is the current stage
-                            current_column_id = template_task.id
-                            all_completed = False
-                            break
-                    else:
-                        # Task doesn't exist yet - this is the current stage
-                        current_column_id = template_task.id
-                        all_completed = False
-                        break
-                
-                if all_completed:
-                    # All template tasks are completed
-                    projects_by_column['completed'].append(project)
-                    project_counts['completed'] += 1
-                elif current_column_id:
-                    projects_by_column[current_column_id].append(project)
-                    project_counts[current_column_id] += 1
-                else:
-                    # Fallback to first column
-                    if kanban_columns:
-                        first_column = kanban_columns[0]
-                        projects_by_column[first_column.id].append(project)
-                        project_counts[first_column.id] += 1
-    else:
-        # If no work type selected, show message to select work type
-        projects_by_column = {}
-        project_counts = {}
-    
-    # Get filter options
-    users = User.query.filter_by(firm_id=firm_id).all()
-    
-    return render_template('projects/kanban_modern.html', 
-                         projects_by_column=projects_by_column,
-                         project_counts=project_counts,
-                         kanban_columns=kanban_columns,
-                         current_work_type=current_work_type,
-                         work_types=work_types,
-                         users=users,
-                         today=date.today())
+# Kanban route moved to views blueprint
 
 # API clients search route moved to api blueprint
         if not new_column_id:
@@ -591,27 +456,7 @@ def kanban_view():
         db.session.rollback()
         return jsonify({'success': False, 'message': str(e)}), 500
 
-@app.route('/api/project/<int:project_id>/progress', methods=['GET'])
-def get_project_progress(project_id):
-    """Get current progress data for a project"""
-    project = Project.query.get_or_404(project_id)
-    
-    # Check access
-    if project.firm_id != session['firm_id']:
-        return jsonify({'success': False, 'message': 'Access denied'}), 403
-    
-    try:
-        # Use Project model's progress calculation (single source of truth)
-        project_tasks = Task.query.filter_by(project_id=project.id).all()
-        
-        return jsonify({
-            'success': True,
-            'progress_percentage': project.progress_percentage,
-            'completed_tasks': len([t for t in project_tasks if t.status == 'Completed']),
-            'total_tasks': len(project_tasks)
-        })
-    except Exception as e:
-        return jsonify({'success': False, 'message': str(e)}), 500
+# API project progress route moved to api blueprint
 
 # Search route moved to views blueprint
     if search_type in ['all', 'clients']:
@@ -635,169 +480,8 @@ def get_project_progress(project_id):
 # Additional client routes moved to clients blueprint
 
 # Admin work_types routes moved to admin blueprint
-@app.route('/admin/work_types/<int:work_type_id>/edit', methods=['POST'])
-def admin_edit_work_type(work_type_id):
-    if session.get('user_role') != 'Admin':
-        return jsonify({'error': 'Access denied'}), 403
-    
-    work_type = WorkType.query.filter_by(id=work_type_id, firm_id=session['firm_id']).first()
-    if not work_type:
-        return jsonify({'error': 'Work type not found'}), 404
-    
-    name = request.form.get('name')
-    color = request.form.get('color')
-    
-    if not name:
-        return jsonify({'error': 'Work type name is required'}), 400
-    
-    # Check if name conflicts with another work type
-    existing = WorkType.query.filter_by(firm_id=session['firm_id'], name=name).filter(WorkType.id != work_type_id).first()
-    if existing:
-        return jsonify({'error': 'Work type with this name already exists'}), 400
-    
-    old_name = work_type.name
-    work_type.name = name
-    if color:
-        work_type.color = color
-    
-    db.session.commit()
-    
-    create_activity_log(f'Work type "{old_name}" updated to "{name}"', session['user_id'])
-    
-    return jsonify({'success': True})
-
-@app.route('/admin/work_types/<int:work_type_id>/statuses/create', methods=['POST'])
-def admin_create_status(work_type_id):
-    if session.get('user_role') != 'Admin':
-        return jsonify({'error': 'Access denied'}), 403
-    
-    work_type = WorkType.query.filter_by(id=work_type_id, firm_id=session['firm_id']).first()
-    if not work_type:
-        return jsonify({'error': 'Work type not found'}), 404
-    
-    name = request.form.get('name')
-    color = request.form.get('color', '#6b7280')
-    is_terminal = request.form.get('is_terminal') == 'true'
-    is_default = request.form.get('is_default') == 'true'
-    
-    if not name:
-        return jsonify({'error': 'Status name is required'}), 400
-    
-    # Check if status already exists for this work type
-    existing = TaskStatus.query.filter_by(work_type_id=work_type_id, name=name).first()
-    if existing:
-        return jsonify({'error': 'Status with this name already exists for this work type'}), 400
-    
-    # Get next position
-    max_position = db.session.query(db.func.max(TaskStatus.position)).filter_by(work_type_id=work_type_id).scalar() or 0
-    
-    # If setting as default, remove default from others
-    if is_default:
-        TaskStatus.query.filter_by(work_type_id=work_type_id, is_default=True).update({'is_default': False})
-    
-    status = TaskStatus(
-        work_type_id=work_type_id,
-        name=name,
-        color=color,
-        position=max_position + 1,
-        is_default=is_default,
-        is_terminal=is_terminal
-    )
-    
-    db.session.add(status)
-    db.session.commit()
-    
-    create_activity_log(f'Status "{name}" created for work type "{work_type.name}"', session['user_id'])
-    
-    return jsonify({'success': True, 'status_id': status.id})
-
-@app.route('/admin/statuses/<int:status_id>/edit', methods=['POST'])
-def admin_edit_status(status_id):
-    if session.get('user_role') != 'Admin':
-        return jsonify({'error': 'Access denied'}), 403
-    
-    status = TaskStatus.query.join(WorkType).filter(
-        TaskStatus.id == status_id,
-        WorkType.firm_id == session['firm_id']
-    ).first()
-    
-    if not status:
-        return jsonify({'error': 'Status not found'}), 404
-    
-    name = request.form.get('name')
-    color = request.form.get('color')
-    is_terminal = request.form.get('is_terminal') == 'true'
-    is_default = request.form.get('is_default') == 'true'
-    
-    if not name:
-        return jsonify({'error': 'Status name is required'}), 400
-    
-    # Check if name conflicts with another status in same work type
-    existing = TaskStatus.query.filter_by(work_type_id=status.work_type_id, name=name).filter(TaskStatus.id != status_id).first()
-    if existing:
-        return jsonify({'error': 'Status with this name already exists for this work type'}), 400
-    
-    # If setting as default, remove default from others
-    if is_default and not status.is_default:
-        TaskStatus.query.filter_by(work_type_id=status.work_type_id, is_default=True).update({'is_default': False})
-    
-    old_name = status.name
-    status.name = name
-    if color:
-        status.color = color
-    status.is_terminal = is_terminal
-    status.is_default = is_default
-    
-    db.session.commit()
-    
-    create_activity_log(f'Status "{old_name}" updated to "{name}" for work type "{status.work_type.name}"', session['user_id'])
-    
-    return jsonify({'success': True})
-
-@app.route('/admin/statuses/<int:status_id>/delete', methods=['POST'])
-def admin_delete_status(status_id):
-    if session.get('user_role') != 'Admin':
-        return jsonify({'error': 'Access denied'}), 403
-    
-    status = TaskStatus.query.join(WorkType).filter(
-        TaskStatus.id == status_id,
-        WorkType.firm_id == session['firm_id']
-    ).first()
-    
-    if not status:
-        return jsonify({'error': 'Status not found'}), 404
-    
-    # Check if status is in use
-    task_count = Task.query.filter_by(work_type_status_id=status_id).count()
-    if task_count > 0:
-        return jsonify({'error': f'Cannot delete status. It is currently used by {task_count} task(s).'}), 400
-    
-    # Check if it's the only status for the work type
-    status_count = TaskStatus.query.filter_by(work_type_id=status.work_type_id).count()
-    if status_count <= 1:
-        return jsonify({'error': 'Cannot delete the only status for a work type'}), 400
-    
-    # If deleting default status, make another one default
-    if status.is_default:
-        next_default = TaskStatus.query.filter_by(work_type_id=status.work_type_id).filter(TaskStatus.id != status_id).first()
-        if next_default:
-            next_default.is_default = True
-    
-    status_name = status.name
-    work_type_name = status.work_type.name
-    
-    db.session.delete(status)
-    db.session.commit()
-    
-    create_activity_log(f'Status "{status_name}" deleted from work type "{work_type_name}"', session['user_id'])
-    
-    return jsonify({'success': True})
-
-@app.route('/admin/work_types/<int:work_type_id>/statuses/reorder', methods=['POST'])
-def admin_reorder_statuses(work_type_id):
-    if session.get('user_role') != 'Admin':
-        return jsonify({'error': 'Access denied'}), 403
-    
+# Admin work_types edit and status create routes moved to admin blueprint
+# Remaining admin status routes moved to admin blueprint
     work_type = WorkType.query.filter_by(id=work_type_id, firm_id=session['firm_id']).first()
     if not work_type:
         return jsonify({'error': 'Work type not found'}), 404
@@ -815,483 +499,18 @@ def admin_reorder_statuses(work_type_id):
     
     return jsonify({'success': True})
 
-# Basic contacts routes moved to contacts blueprint
-@app.route('/contacts/<int:id>')
-def view_contact(id):
-    contact = Contact.query.get_or_404(id)
-    
-    # Get client relationships associated with this contact in current firm
-    client_relationships = db.session.query(ClientContact).join(Client).filter(
-        ClientContact.contact_id == id,
-        Client.firm_id == session['firm_id']
-    ).all()
-    
-    # Get available clients (not already associated with this contact)
-    associated_client_ids = db.session.query(ClientContact.client_id).filter_by(contact_id=id).subquery()
-    available_clients = Client.query.filter(
-        Client.firm_id == session['firm_id'],
-        ~Client.id.in_(associated_client_ids)
-    ).all()
-    
-    return render_template('clients/view_contact.html', contact=contact, client_relationships=client_relationships, available_clients=available_clients)
+# Contact routes moved to contacts blueprint
 
-@app.route('/contacts/<int:id>/edit', methods=['GET', 'POST'])
-def edit_contact(id):
-    contact = Contact.query.get_or_404(id)
-    
-    if request.method == 'POST':
-        contact.first_name = request.form.get('first_name')
-        contact.last_name = request.form.get('last_name')
-        contact.email = request.form.get('email')
-        contact.phone = request.form.get('phone')
-        contact.title = request.form.get('title')
-        contact.company = request.form.get('company')
-        contact.address = request.form.get('address')
-        contact.notes = request.form.get('notes')
-        
-        try:
-            db.session.commit()
-            create_activity_log(f'Contact "{contact.first_name} {contact.last_name}" updated', session['user_id'])
-            flash('Contact updated successfully!', 'success')
-            return redirect(url_for('view_contact', id=contact.id))
-            
-        except Exception as e:
-            db.session.rollback()
-            if 'UNIQUE constraint failed' in str(e):
-                flash('A contact with this email already exists.', 'error')
-            else:
-                flash('Error updating contact.', 'error')
-    
-    return render_template('clients/edit_contact.html', contact=contact)
+# Client and contact association routes moved to clients and contacts blueprints
 
-@app.route('/contacts/<int:contact_id>/clients/<int:client_id>/associate', methods=['POST'])
-def associate_contact_client(contact_id, client_id):
-    contact = Contact.query.get_or_404(contact_id)
-    client = Client.query.get_or_404(client_id)
-    
-    if client.firm_id != session['firm_id']:
-        flash('Access denied', 'error')
-        return redirect(url_for('contacts'))
-    
-    # Check if association already exists
-    existing = ClientContact.query.filter_by(client_id=client_id, contact_id=contact_id).first()
-    if existing:
-        flash('Contact is already associated with this client.', 'error')
-        return redirect(url_for('view_contact', id=contact_id))
-    
-    client_contact = ClientContact(client_id=client_id, contact_id=contact_id)
-    db.session.add(client_contact)
-    db.session.commit()
-    
-    create_activity_log(f'Contact "{contact.first_name} {contact.last_name}" associated with client "{client.name}"', session['user_id'])
-    flash('Contact associated with client successfully!', 'success')
-    
-    return redirect(url_for('view_contact', id=contact_id))
-
-@app.route('/contacts/<int:contact_id>/clients/<int:client_id>/disassociate', methods=['POST'])
-def disassociate_contact_client(contact_id, client_id):
-    client = Client.query.get_or_404(client_id)
-    
-    if client.firm_id != session['firm_id']:
-        flash('Access denied', 'error')
-        return redirect(url_for('contacts'))
-    
-    client_contact = ClientContact.query.filter_by(client_id=client_id, contact_id=contact_id).first()
-    if client_contact:
-        db.session.delete(client_contact)
-        db.session.commit()
-        
-        contact = Contact.query.get(contact_id)
-        create_activity_log(f'Contact "{contact.first_name} {contact.last_name}" disassociated from client "{client.name}"', session['user_id'])
-        flash('Contact disassociated from client successfully!', 'success')
-    
-    return redirect(url_for('view_contact', id=contact_id))
-
-@app.route('/clients/<int:client_id>/associate_contact', methods=['POST'])
-def associate_client_contact(client_id):
-    client = Client.query.get_or_404(client_id)
-    
-    if client.firm_id != session['firm_id']:
-        flash('Access denied', 'error')
-        return redirect(url_for('clients'))
-    
-    contact_id = request.form.get('contact_id')
-    relationship_type = request.form.get('relationship_type')
-    is_primary = request.form.get('is_primary') == '1'
-    
-    if not contact_id:
-        flash('Please select a contact', 'error')
-        return redirect(url_for('view_client', id=client_id))
-    
-    # Check if association already exists
-    existing = ClientContact.query.filter_by(client_id=client_id, contact_id=contact_id).first()
-    if existing:
-        flash('Contact is already associated with this client', 'error')
-        return redirect(url_for('view_client', id=client_id))
-    
-    # If setting as primary, remove primary status from other contacts
-    if is_primary:
-        ClientContact.query.filter_by(client_id=client_id, is_primary=True).update({'is_primary': False})
-    
-    # Create new association
-    client_contact = ClientContact(
-        client_id=client_id,
-        contact_id=contact_id,
-        relationship_type=relationship_type,
-        is_primary=is_primary
-    )
-    
-    db.session.add(client_contact)
-    db.session.commit()
-    
-    contact = Contact.query.get(contact_id)
-    create_activity_log(f'Contact "{contact.full_name}" linked to client "{client.name}" as {relationship_type}', session['user_id'])
-    flash('Contact linked successfully!', 'success')
-    
-    return redirect(url_for('view_client', id=client_id))
-
-@app.route('/contacts/<int:contact_id>/link_client', methods=['POST'])
-def link_contact_client(contact_id):
-    contact = Contact.query.get_or_404(contact_id)
-    
-    client_id = request.form.get('client_id')
-    relationship_type = request.form.get('relationship_type')
-    is_primary = request.form.get('is_primary') == '1'
-    
-    if not client_id:
-        flash('Please select a client', 'error')
-        return redirect(url_for('view_contact', id=contact_id))
-    
-    client = Client.query.get_or_404(client_id)
-    
-    if client.firm_id != session['firm_id']:
-        flash('Access denied', 'error')
-        return redirect(url_for('contacts'))
-    
-    # Check if association already exists
-    existing = ClientContact.query.filter_by(client_id=client_id, contact_id=contact_id).first()
-    if existing:
-        flash('Contact is already associated with this client', 'error')
-        return redirect(url_for('view_contact', id=contact_id))
-    
-    # If setting as primary, remove primary status from other contacts for this client
-    if is_primary:
-        ClientContact.query.filter_by(client_id=client_id, is_primary=True).update({'is_primary': False})
-    
-    # Create new association
-    client_contact = ClientContact(
-        client_id=client_id,
-        contact_id=contact_id,
-        relationship_type=relationship_type,
-        is_primary=is_primary
-    )
-    
-    db.session.add(client_contact)
-    db.session.commit()
-    
-    create_activity_log(f'Contact "{contact.full_name}" linked to client "{client.name}" as {relationship_type}', session['user_id'])
-    flash('Client linked successfully!', 'success')
-    
-    return redirect(url_for('view_contact', id=contact_id))
-
-@app.route('/api/clients')
-def api_clients():
-    """API endpoint to get clients for the current firm"""
-    firm_id = session['firm_id']
-    clients = Client.query.filter_by(firm_id=firm_id, is_active=True).all()
-    
-    return jsonify([{
-        'id': client.id,
-        'name': client.name,
-        'entity_type': client.entity_type
-    } for client in clients])
+# API clients route moved to api blueprint
 
 
-@app.route('/upload', methods=['POST'])
-def upload_file():
-    """Handle file uploads for tasks and projects"""
-    if 'file' not in request.files:
-        return jsonify({'success': False, 'message': 'No file provided'}), 400
-    
-    file = request.files['file']
-    if file.filename == '':
-        return jsonify({'success': False, 'message': 'No file selected'}), 400
-    
-    entity_type = request.form.get('entity_type')  # 'task' or 'project'
-    entity_id = request.form.get('entity_id')
-    
-    if not entity_type or not entity_id:
-        return jsonify({'success': False, 'message': 'Missing entity information'}), 400
-    
-    firm_id = session['firm_id']
-    
-    # Verify entity belongs to firm
-    if entity_type == 'task':
-        task = Task.query.get_or_404(entity_id)
-        if task.project and task.project.firm_id != firm_id:
-            return jsonify({'success': False, 'message': 'Access denied'}), 403
-        elif not task.project and task.firm_id != firm_id:
-            return jsonify({'success': False, 'message': 'Access denied'}), 403
-    elif entity_type == 'project':
-        project = Project.query.get_or_404(entity_id)
-        if project.firm_id != firm_id:
-            return jsonify({'success': False, 'message': 'Access denied'}), 403
-    else:
-        return jsonify({'success': False, 'message': 'Invalid entity type'}), 400
-    
-    # Save file
-    attachment, error = save_uploaded_file(file, firm_id, entity_type, entity_id)
-    
-    if error:
-        return jsonify({'success': False, 'message': error}), 400
-    
-    # Activity log
-    entity_name = task.title if entity_type == 'task' else project.name
-    create_activity_log(
-        f'File "{attachment.original_filename}" uploaded to {entity_type} "{entity_name}"',
-        session['user_id'],
-        project.id if entity_type == 'project' else (task.project_id if task.project else None),
-        task.id if entity_type == 'task' else None
-    )
-    
-    return jsonify({
-        'success': True,
-        'attachment': {
-            'id': attachment.id,
-            'original_filename': attachment.original_filename,
-            'file_size_formatted': attachment.file_size_formatted,
-            'uploaded_at': attachment.uploaded_at.strftime('%m/%d/%Y %I:%M %p'),
-            'uploader_name': attachment.uploader.name
-        }
-    })
+# File upload and attachment routes moved to attachments blueprint
 
-@app.route('/attachments/<int:attachment_id>/download')
-def download_attachment(attachment_id):
-    """Download an attachment"""
-    attachment = Attachment.query.get_or_404(attachment_id)
-    
-    # Check access
-    if attachment.firm_id != session['firm_id']:
-        flash('Access denied', 'error')
-        return redirect(url_for('dashboard.main'))
-    
-    if not os.path.exists(attachment.file_path):
-        flash('File not found', 'error')
-        return redirect(request.referrer or url_for('dashboard.main'))
-    
-    # Activity log
-    entity_type = 'task' if attachment.task_id else 'project'
-    entity_name = attachment.task.title if attachment.task_id else attachment.project.name
-    create_activity_log(
-        f'File "{attachment.original_filename}" downloaded from {entity_type} "{entity_name}"',
-        session['user_id'],
-        attachment.project_id if attachment.project_id else (attachment.task.project_id if attachment.task and attachment.task.project else None),
-        attachment.task_id
-    )
-    
-    return send_file(
-        attachment.file_path,
-        as_attachment=True,
-        download_name=attachment.original_filename,
-        mimetype=attachment.mime_type
-    )
+# Subtask management routes moved to subtasks blueprint
 
-@app.route('/attachments/<int:attachment_id>/delete', methods=['POST'])
-def delete_attachment(attachment_id):
-    """Delete an attachment"""
-    attachment = Attachment.query.get_or_404(attachment_id)
-    
-    # Check access
-    if attachment.firm_id != session['firm_id']:
-        return jsonify({'success': False, 'message': 'Access denied'}), 403
-    
-    try:
-        # Delete physical file
-        if os.path.exists(attachment.file_path):
-            os.remove(attachment.file_path)
-        
-        # Activity log
-        entity_type = 'task' if attachment.task_id else 'project'
-        entity_name = attachment.task.title if attachment.task_id else attachment.project.name
-        create_activity_log(
-            f'File "{attachment.original_filename}" deleted from {entity_type} "{entity_name}"',
-            session['user_id'],
-            attachment.project_id if attachment.project_id else (attachment.task.project_id if attachment.task and attachment.task.project else None),
-            attachment.task_id
-        )
-        
-        # Delete database record
-        db.session.delete(attachment)
-        db.session.commit()
-        
-        return jsonify({'success': True, 'message': 'File deleted successfully'})
-        
-    except Exception as e:
-        db.session.rollback()
-        return jsonify({'success': False, 'message': str(e)}), 500
-
-@app.route('/tasks/<int:task_id>/subtasks/create', methods=['POST'])
-def create_subtask(task_id):
-    """Create a new subtask for a parent task"""
-    parent_task = Task.query.get_or_404(task_id)
-    
-    # Check access
-    if parent_task.project and parent_task.project.firm_id != session['firm_id']:
-        return jsonify({'success': False, 'message': 'Access denied'}), 403
-    elif not parent_task.project and parent_task.firm_id != session['firm_id']:
-        return jsonify({'success': False, 'message': 'Access denied'}), 403
-    
-    try:
-        title = request.json.get('title', '').strip()
-        description = request.json.get('description', '').strip()
-        
-        if not title:
-            return jsonify({'success': False, 'message': 'Title is required'}), 400
-        
-        # Get next subtask order
-        max_order = db.session.query(db.func.max(Task.subtask_order)).filter_by(parent_task_id=task_id).scalar() or 0
-        
-        # Create subtask
-        subtask = Task(
-            title=title,
-            description=description,
-            parent_task_id=task_id,
-            subtask_order=max_order + 1,
-            project_id=parent_task.project_id,
-            firm_id=parent_task.firm_id or session['firm_id'],
-            assignee_id=parent_task.assignee_id,  # Default to parent's assignee
-            priority=parent_task.priority,  # Inherit priority
-            status_id=parent_task.status_id,  # Inherit status system
-            status=parent_task.status if not parent_task.status_id else 'Not Started'
-        )
-        
-        db.session.add(subtask)
-        db.session.commit()
-        
-        # Activity log
-        create_activity_log(
-            f'Subtask "{title}" created for task "{parent_task.title}"',
-            session['user_id'],
-            parent_task.project_id,
-            parent_task.id
-        )
-        
-        return jsonify({
-            'success': True,
-            'subtask': {
-                'id': subtask.id,
-                'title': subtask.title,
-                'description': subtask.description,
-                'status': subtask.current_status,
-                'assignee_name': subtask.assignee.name if subtask.assignee else 'Unassigned',
-                'created_at': subtask.created_at.strftime('%m/%d/%Y %I:%M %p')
-            }
-        })
-        
-    except Exception as e:
-        db.session.rollback()
-        return jsonify({'success': False, 'message': str(e)}), 500
-
-@app.route('/tasks/<int:task_id>/subtasks/reorder', methods=['POST'])
-def reorder_subtasks(task_id):
-    """Reorder subtasks within a parent task"""
-    parent_task = Task.query.get_or_404(task_id)
-    
-    # Check access
-    if parent_task.project and parent_task.project.firm_id != session['firm_id']:
-        return jsonify({'success': False, 'message': 'Access denied'}), 403
-    elif not parent_task.project and parent_task.firm_id != session['firm_id']:
-        return jsonify({'success': False, 'message': 'Access denied'}), 403
-    
-    try:
-        subtask_ids = request.json.get('subtask_ids', [])
-        
-        # Update order for each subtask
-        for index, subtask_id in enumerate(subtask_ids):
-            subtask = Task.query.get(subtask_id)
-            if subtask and subtask.parent_task_id == task_id:
-                subtask.subtask_order = index + 1
-        
-        db.session.commit()
-        
-        return jsonify({'success': True, 'message': 'Subtasks reordered successfully'})
-        
-    except Exception as e:
-        db.session.rollback()
-        return jsonify({'success': False, 'message': str(e)}), 500
-
-@app.route('/tasks/<int:task_id>/convert-to-subtask', methods=['POST'])
-def convert_to_subtask(task_id):
-    """Convert an existing task to a subtask of another task"""
-    task = Task.query.get_or_404(task_id)
-    
-    # Check access
-    if task.project and task.project.firm_id != session['firm_id']:
-        return jsonify({'success': False, 'message': 'Access denied'}), 403
-    elif not task.project and task.firm_id != session['firm_id']:
-        return jsonify({'success': False, 'message': 'Access denied'}), 403
-    
-    try:
-        parent_task_id = request.json.get('parent_task_id')
-        
-        if not parent_task_id:
-            return jsonify({'success': False, 'message': 'Parent task ID is required'}), 400
-        
-        parent_task = Task.query.get(parent_task_id)
-        if not parent_task:
-            return jsonify({'success': False, 'message': 'Parent task not found'}), 404
-        
-        # Check that parent task belongs to same firm
-        if parent_task.project and parent_task.project.firm_id != session['firm_id']:
-            return jsonify({'success': False, 'message': 'Access denied'}), 403
-        elif not parent_task.project and parent_task.firm_id != session['firm_id']:
-            return jsonify({'success': False, 'message': 'Access denied'}), 403
-        
-        # Prevent circular relationships
-        if parent_task_id == task_id:
-            return jsonify({'success': False, 'message': 'Cannot make task a subtask of itself'}), 400
-        
-        # Check if parent task is already a subtask of this task (prevent circular)
-        current = parent_task
-        while current.parent_task:
-            if current.parent_task.id == task_id:
-                return jsonify({'success': False, 'message': 'Cannot create circular subtask relationship'}), 400
-            current = current.parent_task
-        
-        # Get next subtask order
-        max_order = db.session.query(db.func.max(Task.subtask_order)).filter_by(parent_task_id=parent_task_id).scalar() or 0
-        
-        # Convert to subtask
-        task.parent_task_id = parent_task_id
-        task.subtask_order = max_order + 1
-        
-        db.session.commit()
-        
-        # Activity log
-        create_activity_log(
-            f'Task "{task.title}" converted to subtask of "{parent_task.title}"',
-            session['user_id'],
-            parent_task.project_id,
-            parent_task.id
-        )
-        
-        return jsonify({'success': True, 'message': 'Task converted to subtask successfully'})
-        
-    except Exception as e:
-        db.session.rollback()
-        return jsonify({'success': False, 'message': str(e)}), 500
-
-@app.route('/admin/process-recurring', methods=['POST'])
-def admin_process_recurring():
-    """Manual trigger for processing recurring tasks"""
-    if session.get('user_role') != 'Admin':
-        return jsonify({'error': 'Access denied'}), 403
-    
-    try:
-        process_recurring_tasks()
-        return jsonify({'success': True, 'message': 'Recurring tasks processed successfully'})
-    except Exception as e:
-        return jsonify({'success': False, 'message': str(e)})
+# Admin recurring task processing route moved to admin blueprint
 
 # CPA Document Checklist Management Routes
 @app.route('/checklists')
